requirements.txt
Flask==2.3.2
pymongo==4.4.0
gridfs==0.7.1
pandas==2.2.0
numpy==1.26.0
matplotlib==3.8.0
python-dotenv==1.0.0